Chores/match test organization with main (#1232)
# Goal
The goal of this PR is to reorganize tests to match what @rlaferla did in the recently merged PR to main. Related to #1224 

* Capacity tests, mock, and testing_utils have been moved to a tests module. 
* Tests are split up by what they test.
* Additionally, removed the setup of a runtime environment from some tests
that didn't need it.
