+++ pallets/capacity/src/tests/mock.rs
@@ -1,4 +1,3 @@
-use super::*;
 use crate as pallet_capacity;
 
 use common_primitives::{
@@ -14,7 +13,7 @@ use sp_core::{ConstU8, H256};
 use sp_runtime::{
 	testing::Header,
 	traits::{BlakeTwo256, Convert, IdentityLookup},
-	AccountId32,
+	AccountId32, DispatchError,
 };
 
 type UncheckedExtrinsic = frame_system::mocking::MockUncheckedExtrinsic<Test>;
@@ -152,7 +151,7 @@ impl pallet_msa::Config for Test {
 	type MaxSignaturesStored = ConstU32<8000>;
 }
 
-impl Config for Test {
+impl pallet_capacity::Config for Test {
 	type RuntimeEvent = RuntimeEvent;
 	type WeightInfo = ();
 	type Currency = pallet_balances::Pallet<Self>;
