+++ pallets/capacity/src/tests/mod.rs
@@ -1,0 +1,12 @@
+pub mod capacity_details_tests;
+pub mod epochs_tests;
+pub mod mock;
+pub mod other_tests;
+pub mod replenishment_tests;
+pub mod stake_and_deposit_tests;
+pub mod staking_account_details_tests;
+pub mod staking_target_details_tests;
+pub mod testing_utils;
+pub mod unstaking_tests;
+pub mod withdraw_unstaked_tests;
+pub mod withdrawal_tests;
