+++ pallets/capacity/src/tests/testing_utils.rs
@@ -1,15 +1,17 @@
-use super::*;
-use crate::mock::*;
+use super::mock::*;
+use frame_support::{assert_ok, traits::Hooks};
 
-use frame_support::assert_ok;
+#[allow(unused)]
+use sp_runtime::traits::SignedExtension;
+
+use crate::{BalanceOf, CapacityDetails, Config, Event};
+use common_primitives::msa::MessageSourceId;
 
 pub fn staking_events() -> Vec<Event<Test>> {
 	let result = System::events()
 		.into_iter()
 		.map(|r| r.event)
-		.filter_map(
-			|e| if let mock::RuntimeEvent::Capacity(inner) = e { Some(inner) } else { None },
-		)
+		.filter_map(|e| if let RuntimeEvent::Capacity(inner) = e { Some(inner) } else { None })
 		.collect::<Vec<_>>();
 
 	System::reset_events();
